chore(backend): rename OpenAPI middleware files
